with entities, repos, db connection id types Long

+++ src/main/java/by/vstu/isit/documentprocessor/entities/Func.java
@@ -41,7 +41,7 @@ public class Func implements Serializable {
     private String name;
 
     @Column(name = "idOper", nullable = false)
-    private Integer idOper;
+    private Long idOper;
 
     /**
      * ПУ-8
